refactor(log): Update namespace from `logger` to `log` for consistency across log targets.
